+++ packages/headless/src/features/fields/fields-slice.ts
@@ -1,6 +1,11 @@
 import {createReducer} from '@reduxjs/toolkit';
 import {getFieldsInitialState} from './fields-state';
-import {registerFieldsToInclude} from './fields-actions';
+import {
+  disableFetchAllFields,
+  enableFetchAllFields,
+  fetchFieldsDescription,
+  registerFieldsToInclude,
+} from './fields-actions';
 import {registerFolding} from '../folding/folding-actions';
 import {getFoldingInitialState} from '../folding/folding-state';
 
@@ -10,6 +15,15 @@ export const fieldsReducer = createReducer(getFieldsInitialState(), (builder) =>
       state.fieldsToInclude = [
         ...new Set(state.fieldsToInclude.concat(action.payload)),
       ];
+    })
+    .addCase(enableFetchAllFields, (state) => {
+      state.fetchAllFields = true;
+    })
+    .addCase(disableFetchAllFields, (state) => {
+      state.fetchAllFields = false;
+    })
+    .addCase(fetchFieldsDescription.fulfilled, (state, {payload}) => {
+      state.fieldsDescription = payload;
     })
     .addCase(registerFolding, (state, {payload}) => {
       const defaultFields = getFoldingInitialState().fields;