+++ packages/quantic/force-app/main/default/lwc/quanticDateFacet/quanticDateFacet.js
@@ -3,7 +3,7 @@ import {
   registerComponentForInit,
   initializeWithHeadless,
 } from 'c/quanticHeadlessLoader';
-import {I18nUtils} from 'c/quanticUtils';
+import {I18nUtils, fromSearchApiDate} from 'c/quanticUtils';
 import LOCALE from '@salesforce/i18n/locale';
 
 import clearFilter from '@salesforce/label/c.quantic_ClearFilter';
@@ -63,9 +63,9 @@ export default class QuanticDateFacet extends LightningElement {
    * @returns {string}
    */
   @api formattingFunction = (item) => `${new Intl.DateTimeFormat(LOCALE).format(
-    new Date(item.start)
+    new Date(fromSearchApiDate(item.start))
   )} - ${new Intl.DateTimeFormat(LOCALE).format(
-    new Date(item.end)
+    new Date(fromSearchApiDate(item.end))
   )}`;
   /** 
    * Whether the facet is collapsed.