+++ packages/quantic/force-app/main/default/lwc/quanticUtils/quanticUtils.js
@@ -155,3 +155,16 @@ export function setItemInLocalStorage(key, item) {
 export function regexEncode(value) {
   return value.replace(/[-[\]/{}()*+?.\\^$|]/g, '\\$&');
 }
+
+/**
+ * Converts a date string from the Coveo Search API format to the ISO-8601 format.
+ * Replace `/` characters in date string with `-`.
+ * Replace `@` characters in date string with `T`.
+ * @param {string} dateString 
+ * @returns {string}
+ */
+export function fromSearchApiDate(dateString) {
+  return dateString
+    .replaceAll('/', '-')
+    .replaceAll('@', 'T');
+}